+++ src/BabylonCpp/src/mesh/lines_mesh.cpp
@@ -112,7 +112,8 @@ void LinesMesh::_draw(SubMesh* subMesh, int /*fillMode*/,
   auto engine = getScene()->getEngine();
 
   // Draw order
-  engine->draw(false, subMesh->indexStart, subMesh->indexCount);
+  engine->draw(false, subMesh->indexStart,
+               static_cast<int>(subMesh->indexCount));
 }
 
 PickingInfo LinesMesh::intersects(const Ray& /*ray*/, bool /*fastCheck*/)
